+++ src/renderer/entities/staking/api/__tests__/stakingDataService.test.ts
