+++ src/renderer/entities/staking/api/__tests__/stakingRewardsService.test.ts
