+++ src/renderer/entities/staking/api/apyCalculator.ts
@@ -1,8 +1,8 @@
 import { ApiPromise } from '@polkadot/api';
 import BigNumber from 'bignumber.js';
 
-import { ApyValidator } from './common/types';
-import { DECAY_RATE, INTEREST_IDEAL, MINIMUM_INFLATION, STAKED_PORTION_IDEAL } from './common/constants';
+import { ApyValidator } from '../lib/types';
+import { DECAY_RATE, INTEREST_IDEAL, MINIMUM_INFLATION, STAKED_PORTION_IDEAL } from '../lib/constants';
 import type { Address } from '@shared/core';
 
 const calculateYearlyInflation = (stakedPortion: number): number => {
