+++ src/renderer/entities/staking/api/stakingDataService.ts
@@ -1,6 +1,6 @@
 import { ApiPromise } from '@polkadot/api';
 
-import { IStakingDataService, StakingMap } from './common/types';
+import { IStakingDataService, StakingMap } from '../lib/types';
 import type { Address, ChainId, EraIndex, Unlocking } from '@shared/core';
 import { redeemableAmount, ZERO_BALANCE } from '@shared/lib/utils';
