+++ src/renderer/entities/staking/graphql/queries/stakingRewards.ts
