+++ src/renderer/entities/staking/graphql/types/stakingRewards.ts
