+++ src/renderer/entities/staking/index.ts
@@ -1,2 +1,5 @@
-export * from './lib';
 export * from './ui';
+export * from './api';
+export * from './lib/types';
+export * from './lib/constants';
+export { useValidatorsMap } from './lib/useValidatorsMap';
