--- src/renderer/entities/staking/lib/index.ts
@@ -1,8 +1,0 @@
-export * from './apyCalculator';
-export * from './eraService';
-export * from './stakingDataService';
-export * from './stakingRewardsService';
-export { validatorsService } from './validatorsService';
-export * from './common/types';
-export * from './common/constants';
-export { useValidatorsMap } from './useValidatorsMap';
