+++ src/renderer/entities/staking/lib/types.ts
