+++ src/renderer/entities/staking/lib/useValidatorsMap.ts
@@ -1,7 +1,8 @@
 import { useEffect, useState } from 'react';
 import { ApiPromise } from '@polkadot/api';
 
-import { ValidatorMap, useEra, validatorsService } from './';
+import { ValidatorMap } from './types';
+import { useEra, validatorsService } from '../api';
 
 export const useValidatorsMap = (api?: ApiPromise, isLightClient?: boolean): ValidatorMap => {
   const { subscribeActiveEra } = useEra();
